// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserStatus = {
  INACTIVE: 'INACTIVE',
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  BLOCKED: 'BLOCKED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const UserRole = {
  INVESTOR: 'INVESTOR',
  TEAM_LEADER: 'TEAM_LEADER',
  FINANCE_MANAGER: 'FINANCE_MANAGER',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const TransactionType = {
  REGISTRATION_FEE: 'REGISTRATION_FEE',
  WALLET_TOPUP: 'WALLET_TOPUP',
  INVESTMENT_DEBIT: 'INVESTMENT_DEBIT',
  PROFIT_CREDIT: 'PROFIT_CREDIT',
  WITHDRAWAL: 'WITHDRAWAL',
  REFUND: 'REFUND',
  ADMIN_CREDIT: 'ADMIN_CREDIT',
  ADMIN_DEBIT: 'ADMIN_DEBIT'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const TransactionStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  REVERSED: 'REVERSED'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]


export const ProjectStatus = {
  PLANNED: 'PLANNED',
  LAND_ACQUIRED: 'LAND_ACQUIRED',
  LEGAL_IN_PROGRESS: 'LEGAL_IN_PROGRESS',
  APPROVED: 'APPROVED',
  UNDER_DEVELOPMENT: 'UNDER_DEVELOPMENT',
  READY_FOR_SALE: 'READY_FOR_SALE',
  COMPLETED: 'COMPLETED',
  CLOSED: 'CLOSED'
} as const

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus]


export const PlotStatus = {
  AVAILABLE: 'AVAILABLE',
  HELD: 'HELD',
  RESERVED: 'RESERVED',
  SOLD: 'SOLD'
} as const

export type PlotStatus = (typeof PlotStatus)[keyof typeof PlotStatus]


export const PlotType = {
  CORNER: 'CORNER',
  MIDDLE: 'MIDDLE',
  DOUBLE_ROAD: 'DOUBLE_ROAD'
} as const

export type PlotType = (typeof PlotType)[keyof typeof PlotType]


export const SalePaymentStatus = {
  PENDING: 'PENDING',
  PARTIAL: 'PARTIAL',
  COMPLETED: 'COMPLETED'
} as const

export type SalePaymentStatus = (typeof SalePaymentStatus)[keyof typeof SalePaymentStatus]


export const HoldStatus = {
  ACTIVE: 'ACTIVE',
  RELEASED_MANUAL: 'RELEASED_MANUAL',
  RELEASED_EXPIRED: 'RELEASED_EXPIRED',
  RELEASED_ADMIN: 'RELEASED_ADMIN',
  CONVERTED_TO_SALE: 'CONVERTED_TO_SALE'
} as const

export type HoldStatus = (typeof HoldStatus)[keyof typeof HoldStatus]


export const DistributionStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  ROLLED_BACK: 'ROLLED_BACK'
} as const

export type DistributionStatus = (typeof DistributionStatus)[keyof typeof DistributionStatus]